update signature contract, chainId use uint64

+++ chains/evm/cli/admin/flagVars.go
@@ -14,7 +14,7 @@ var (
 	Relayer            string
 	DepositNonce       uint64
 	DomainID           uint8
-	ChainID            uint8
+	ChainID            uint64
 	Fee                string
 	RelayerThreshold   uint64
 	SignatureThreshold uint64
